delete folders, recreate personnel database
