--- hw05/Employee Search/Employee Search.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.7.34031.279
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Employee Search", "Employee Search.csproj", "{BFF2AC9C-A651-421B-AE8E-7752E2DC2B19}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{BFF2AC9C-A651-421B-AE8E-7752E2DC2B19}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BFF2AC9C-A651-421B-AE8E-7752E2DC2B19}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BFF2AC9C-A651-421B-AE8E-7752E2DC2B19}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BFF2AC9C-A651-421B-AE8E-7752E2DC2B19}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {CF76EEB1-F519-402D-AD94-F6B05DDD898A}
-	EndGlobalSection
-EndGlobal
