+++ hw05/Personel Database/Personel Database.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.7.34031.279
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Personnel Database", "Personnel Database\Personnel Database.csproj", "{26A25F43-B11F-4875-954D-2F16291AB95B}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{26A25F43-B11F-4875-954D-2F16291AB95B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{26A25F43-B11F-4875-954D-2F16291AB95B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{26A25F43-B11F-4875-954D-2F16291AB95B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{26A25F43-B11F-4875-954D-2F16291AB95B}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {F79466B5-7D8F-4E65-81DF-7181122AFD06}
+	EndGlobalSection
+EndGlobal
