--- hw05/Personnel Database/App.config
@@ -1,13 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<configuration>
-    <configSections>
-    </configSections>
-    <connectionStrings>
-        <add name="Personnel_Database.Properties.Settings.PersonnelConnectionString"
-            connectionString="Data Source=(LocalDB)\MSSQLLocalDB;AttachDbFilename=|DataDirectory|\Personnel.mdf;Integrated Security=True"
-            providerName="System.Data.SqlClient" />
-    </connectionStrings>
-    <startup> 
-        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.7.2" />
-    </startup>
-</configuration>
