--- hw05/Personnel Database/Personnel Database.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.7.34031.279
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Personnel Database", "Personnel Database.csproj", "{10585A37-BFD9-4BCA-A364-EAB106784488}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{10585A37-BFD9-4BCA-A364-EAB106784488}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{10585A37-BFD9-4BCA-A364-EAB106784488}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{10585A37-BFD9-4BCA-A364-EAB106784488}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{10585A37-BFD9-4BCA-A364-EAB106784488}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {96DC2C1B-E058-4820-A483-CF89569DBC0B}
-	EndGlobalSection
-EndGlobal
